Remove rest of int32 from the Main Libraries

+++ Libraries/GPT/GPT/Main.go
@@ -42,7 +42,7 @@ parameter.
 – Returns:
   - the entry or an empty entry with time = -1 if it doesn't exist
  */
-func GetEntry(time int64, num int32) *Entry {
+func GetEntry(time int64, num int) *Entry {
 	var page_contents string = string(Utils.GetPageContentsWEBSITE(Utils.PersonalConsts_GL.WEBSITE_URL +
 		"files_EOG/gpt_text.txt", Utils.PersonalConsts_GL.WEBSITE_PW))
 	if page_contents == "" {
@@ -78,9 +78,9 @@ func GetEntry(time int64, num int32) *Entry {
 		}
 
 		if num < 0 {
-			num = int32(len(entries)) + num
-		} else if num >= int32(len(entries)) {
-			num = int32(len(entries)) - 1
+			num = len(entries) + num
+		} else if num >= len(entries) {
+			num = len(entries) - 1
 		} else {
 			// Do nothing
 		}
